Created Sql2oBusinessDaoTest file
